pyproject.toml
[tool.poetry]
name = "threaduler"
version = "0.1.0"
description = ""
authors = ["maikroservice <20245897+maikroservice@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
fastapi = "^0.86.0"
requests = "^2.28.1"
uvicorn = {extras = ["standard"], version = "^0.19.0"}
tweepy = "^4.14.0"
python-magic = "^0.4.27"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
beautifulsoup4 = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
